+++ app/Components/Model/Model.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Solital\Components\Model;
+
+abstract class Model
+{
+    protected $table;
+    protected $columnPrimaryKey;
+    protected $columns;
+}