+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Released Notes
 
+## v0.3.3 - (2020-06-19)
+
+### Added
+
+- Controller and Model folder
+
+### Fixed
+
+- Fixed Vinci constant
+
+## v0.3.2 - (2020-06-19)
+
+### Fixed
+
+- scripts in composer.json
+
+## v0.3.1 - (2020-06-19)
+
+### Fixed
+
+- __DIR__ in index.php
+
 ## v0.3.0 - (2020-06-19)
 
 ### Added